% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solvers.R
\name{check_solver}
\alias{check_solver}
\title{Checks if solver exists}
\usage{
check_solver(solver)
}
\arguments{
\item{solver}{name of the solver (it has to be in the PATH)}
}
\value{
TRUE if the solver exists and FALSE otherwise
}
\description{
Function that checks if the solver exists and it can be run without any issue
}
\examples{
check_solver(solver = "ipopt")

}
