% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solfile.R
\name{example_sol_file}
\alias{example_sol_file}
\title{Returns an example.sol file}
\usage{
example_sol_file()
}
\value{
An example.sol file
}
\description{
Function that returns an example.sol file
}
\examples{
example_sol_file()

}
