% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solvers.R
\name{get_available_options}
\alias{get_available_options}
\title{Gets the available options for a solver}
\usage{
get_available_options(solver)
}
\arguments{
\item{solver}{name of the solver (it has to be in the PATH)}
}
\value{
vector with all the available options for a solver
}
\description{
Function that gets a vector with all the available options for a solver
}
\examples{
get_available_options(solver = "ipopt")

}
