% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batteries.R
\name{get_minlplib_problem}
\alias{get_minlplib_problem}
\title{Gets problem from MINLPLib}
\usage{
get_minlplib_problem(name)
}
\arguments{
\item{name}{name of the problem}
}
\value{
TRUE if there is no error getting the problem and FALSE otherwise
}
\description{
Function that gets the corresponding problem from MINLPLib library
}
\examples{
get_minlplib_problem(name = "alkyl")

}
