% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solvers.R
\name{print_available_options}
\alias{print_available_options}
\title{Prints the available options of the solver}
\usage{
print_available_options(solver)
}
\arguments{
\item{solver}{name of the solver (it has to be in the PATH)}
}
\value{
No return value
}
\description{
Function that prints the available options of the solver
}
\examples{
print_available_options(solver = "ipopt")

}
