% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solvers.R
\name{print_solver_version}
\alias{print_solver_version}
\title{Prints the version of the solver}
\usage{
print_solver_version(solver)
}
\arguments{
\item{solver}{name of the solver (it has to be in the PATH)}
}
\value{
No return value
}
\description{
Function that prints the version of the solver
}
\examples{
print_solver_version(solver = "ipopt")

}
