% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solfile.R
\name{read_sol_file}
\alias{read_sol_file}
\title{Reads the .sol file}
\usage{
read_sol_file(solfile)
}
\arguments{
\item{solfile}{.sol file with the solution}
}
\value{
list with a string of the output given by the solver (output), the optimal value of the problem (objective), the status returned by the solver (status), the optimal primal solution (primal_solution), and the optimal dual solution (dual_solution)
}
\description{
Function that reads the corresponding .sol file
}
\examples{
read_sol_file(solfile = "example.sol")

}
